+++ src/tap.ts
@@ -1,4 +1,4 @@
-export function tap<T>(value: T, fn: (value: Readonly<T>) => any): T {
-  fn(Object.freeze(value));
+export function tap<T>(value: T, fn: (value: T) => any): T {
+  fn(value);
   return value;
 }